PostDto 수정
