--- src/main/java/com/wwg/gabozzago/domain/post/Dto/PostDto.java
@@ -1,16 +1,0 @@
-package com.wwg.gabozzago.domain.post.Dto;
-
-import com.wwg.gabozzago.domain.entity.Post;
-import lombok.*;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-
-public class PostDto {
-    private String title;
-    private String location;
-    private String post_img;
-    private String content;
-}
