+++ src/main/java/com/wwg/gabozzago/domain/post/service/PostService.java
@@ -1,0 +1,7 @@
+package com.wwg.gabozzago.domain.post.service;
+
+import com.wwg.gabozzago.domain.post.dto.PostDto;
+
+public interface PostService {
+    void createPost(PostDto postDto);
+}
